+++ src/lib/Scenes/Artwork/Components/ImageCarousel/FullScreen/boxShadow.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from "react-native"
+
+export const boxShadow = StyleSheet.create({
+  boxShadow: {
+    shadowColor: "black",
+    shadowOffset: { width: 0, height: 1 },
+    shadowOpacity: 0.1,
+  },
+}).boxShadow
