+++ src/lib/Scenes/Artwork/Components/ImageCarousel/FullScreen/screen.ts
@@ -1,0 +1,5 @@
+import { Dimensions } from "react-native"
+
+export const screenWidth = Dimensions.get("screen").width
+export const screenHeight = Dimensions.get("screen").height
+export const screenBoundingBox = { width: screenWidth, height: screenHeight }
