+++ src/lib/Scenes/Artwork/Components/ImageCarousel/ImageCarouselContext.tsx
@@ -1,5 +1,5 @@
 import { observable } from "mobx"
-import React, { useContext, useMemo, useRef } from "react"
+import React, { useMemo, useRef } from "react"
 import { FlatList, View } from "react-native"
 
 export interface ImageDescriptor {
@@ -35,31 +35,32 @@ export interface ImageCarouselContext {
     isZoomedCompletelyOut: boolean
   }
   images: ImageDescriptor[]
-  baseImageRefs: View[]
-  baseFlatListRef: React.RefObject<FlatList<any>>
+  embeddedImageRefs: View[]
+  embeddedFlatListRef: React.RefObject<FlatList<any>>
   dispatch(action: ImageCarouselAction): void
 }
 
 export function useNewImageCarouselContext({ images }: { images: ImageDescriptor[] }): ImageCarouselContext {
-  const baseImageRefs = useMemo(() => [], [])
-  const baseFlatListRef = useRef<FlatList<any>>()
+  const embeddedImageRefs = useMemo(() => [], [])
+  const embeddedFlatListRef = useRef<FlatList<any>>()
   const state = observable({
     imageIndex: 0,
     fullScreenState: "none" as FullScreenState,
     isZoomedCompletelyOut: true,
   })
+
   return useMemo(
     () => ({
       state,
       images,
-      baseImageRefs,
-      baseFlatListRef,
+      embeddedImageRefs,
+      embeddedFlatListRef,
       dispatch: (action: ImageCarouselAction) => {
         switch (action.type) {
           case "IMAGE_INDEX_CHANGED":
             state.imageIndex = action.nextImageIndex
             if (state.fullScreenState !== "none") {
-              baseFlatListRef.current.scrollToIndex({ index: action.nextImageIndex, animated: false })
+              embeddedFlatListRef.current.scrollToIndex({ index: action.nextImageIndex, animated: false })
             }
             break
           case "FULL_SCREEN_DISMISSED":
@@ -79,14 +80,6 @@ export function useNewImageCarouselContext({ images }: { images: ImageDescriptor
     }),
     []
   )
-}
-
-export const useImageIndex = () => {
-  return useContext(ImageCarouselContext).state.imageIndex
-}
-
-export const useFullScreenState = () => {
-  return useContext(ImageCarouselContext).state.fullScreenState
 }
 
 export const ImageCarouselContext = React.createContext<ImageCarouselContext>(null)
